backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
PyJWT==2.8.0
python-multipart==0.0.12
openpyxl==3.1.5
httpx==0.27.0
pytest==8.3.3
pytest-asyncio==0.24.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
idna==3.11
pydantic_core==2.27.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.15.0
python-dotenv==1.0.0
python-telegram-bot==21.9
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0